change constraint type

+++ Messages/Validation/LoadMessagesValidator.cs
@@ -5,7 +5,7 @@ namespace OnChat.Shared.Messages.Validation;
 
 public class LoadMessagesValidator : ValidatorBase<LoadMessagesPacket, PacketId>
 {
-    public readonly MinMaxConstraint<byte> QuantityConstraint = new(new MinMaxValue<byte>(0, 255));
+    public readonly MinMaxConstraint<int> QuantityConstraint = new(new MinMaxValue<int>(0, 255));
     
     public LoadMessagesValidator()
     {
